+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v0.7.6 (TBD)
  - [Export] PNG support for figure export
  - [Export] Bundle PhantomJS in executables or load from APP_DIR
  - [Misc] Store last positions of certain windows (e.g., UserSettings, ExportFigure) 
+ - [Plots] Grid line customization (color, thickness, alpha)
 
 
 v0.7.5 (2020.04.09)